Fix kwarg updates in from_pretrained & main rebase
