+++ src/transformers/models/timm_backbone/modeling_timm_backbone.py
@@ -16,8 +16,9 @@
 from typing import Tuple, Union
 
 from ...modeling_outputs import BackboneOutput
-from ...modeling_utils import BackboneMixin, PreTrainedModel
+from ...modeling_utils import PreTrainedModel
 from ...utils import is_torch_available, requires_backends
+from ...utils.backbone_utils import BackboneMixin
 from .configuration_timm_backbone import TimmBackboneConfig
